Add pip_lock_down dependency
The repo server, like any other host type in the openstack-ansible
project, needs to be able to support the use of additional software. An
example of this would be some form of logging solution. For this to
work the repo server needs to be using the same packages as other hosts
in the deployment.

This commit adds the pip_lock_down role as a dependency so that from
pip's perspective the repo server can be treated like any other host.
The need for the repo_build process to be able to access packages
outside of what is held by the repo server is moved to that role by the
change referenced below.

Change-Id: Ib53eed946c721bf497c3e83db371e45650f2ad05
Depends-on: I2fbcb76f349242bf45161a1789891ed3133557eb
Depends-on: Ifdb21cfb05ba412e6959cd53d7bdeeaa2e43fae7